requirements.txt
asgiref==3.2.3
Django==3.0.2
django-extensions==2.2.6
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
graphviz==0.13.2
psycopg2==2.8.4
six==1.14.0
sqlparse==0.3.0
